requirements.txt
absl-py==1.4.0
alabaster==0.7.13
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.0.0
astor==0.8.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==4.2.4
charset-normalizer==3.2.0
comm==0.1.4
cycler==0.11.0
Cython==0.29.24
debugpy==1.6.7.post1
Django==2.2.3
django-cors-headers==1.2.2
djangorestframework==3.10.3
djangorestframework-simplejwt==5.2.2
docutils==0.19
ecdsa==0.18.0
exceptiongroup==1.1.2
fastjsonschema==2.18.0
fonttools==4.38.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
gunicorn==19.9.0
h5py==2.10.0
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.16.2
ipyparallel==8.6.1
ipython==7.34.0
ipywidgets==8.1.0
jedi==0.19.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.3
mask-rcnn-tf2==1.0
matplotlib==3.5.3
mccabe==0.6.1
mistune==2.0.4
mypy-extensions==0.4.3
mysqlclient==1.4.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.2.1
nbformat==5.8.0
nest-asyncio==1.5.7
networkx==2.6.3
nose==1.3.7
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.20.3
oauthlib==3.2.2
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==23.0
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
PyJWT==2.4.0
pyls-spyder==0.4.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-jose==3.3.0
python-lsp-jsonrpc==1.0.0
PyWavelets==1.3.0
pywin32==306
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.8.2
shapely==2.0.1
sip==4.19.13
sniffio==1.3.0
soupsieve==2.4.1
Sphinx==5.3.0
sqlparse==0.4.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==2.3.0
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2020.5
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
whitenoise==5.3.0
widgetsnbextension==4.0.8
wincertstore==0.2
wrapt==1.15.0
zipp==3.15.0
